pyproject.toml
[project]
name = "core-llm-bridge"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
    {name = "Fernando Velasco", email = "fermaat.vl@gmail.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = "==3.12.*"
keywords = ["llm", "ollama", "language-model", "bridge", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.25",
    "loguru>=0.7",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "isort>=5.12",
]

docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=1.3",
    "sphinx-autodoc-typehints>=1.23",
]

[project.urls]
Homepage = "https://github.com/yourusername/llm-bridge"
Repository = "https://github.com/yourusername/llm-bridge.git"
Documentation = "https://github.com/yourusername/llm-bridge#readme"
Issues = "https://github.com/yourusername/llm-bridge/issues"

[tool.pdm]
distribution = false

[tool.pdm.build]
includes = ["src/core_llm_bridge"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "--strict-markers -v --cov=src/llm_bridge --cov-report=term-missing --cov-report=html"
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "requires_ollama: tests that require Ollama running",
    "asyncio: asynchronous tests",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"
extend-select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C4",     # flake8-comprehensions
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TRY",    # tryceratops
    "ASYNC",  # flake8-async
]
ignore = [
    "E501",   # line-length (handled by black)
    "E203",   # whitespace before ':'
    "E266",   # too many leading '#' for block comment
    "W605",   # invalid escape sequence
]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".mypy_cache",
    ".tox",
    ".eggs",
]

[tool.ruff.isort]
combine-as-imports = true
force-single-line = false
known-first-party = ["core_llm_bridge"]

[tool.coverage.run]
source = ["src/core_llm_bridge"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
